+++ src/03_RPV_modelling.py
@@ -70,21 +70,21 @@ def main():
 
     # Create rpvs for each
     for week, gdf in weeks_dics.items():
-        out_dir = Path(base_dir) / "RPV_Results" / "V9" / week
+        out_dir = Path(base_dir) / "RPV_Results" / "V11" / week
         out_dir.mkdir(parents=True, exist_ok=True)
-        plot_df(week, gdf, out_dir)
+        #plot_df(week, gdf, out_dir)
 
         for band in bands:
             if (out_dir / f"rpv_{week}_{band}_results.csv").exists():
                 continue
-            result = process_weekly_data({week: gdf}, band=band,sample_total_dataset=500_000)
+            result = process_weekly_data({week: gdf}, band=band,sample_total_dataset=500_000, filter={})
             result.drop("geometry").write_csv(str(out_dir / f"rpv_{week}_{band}_results.csv"))
 
 
     # Create Plots of RPV results
     df_all_rpv = pd.DataFrame()
     for week, gdf in weeks_dics.items():
-        out_dir = Path(base_dir) / "RPV_Results" / "V9" / week
+        out_dir = Path(base_dir) / "RPV_Results" / "V11" / week
         for band in bands:
             df_rpv = pd.read_csv(str(out_dir / f"rpv_{week}_{band}_results.csv"))
             df_rpv["week"] = week
@@ -92,7 +92,7 @@ def main():
             df_all_rpv = pd.concat([df_all_rpv, df_rpv])
     df_all_rpv.reset_index(inplace=True)
 
-    df_all_rpv.to_csv(str(Path(base_dir) / "RPV_Results" / "V9"  / "rpv_results.csv"))
+    df_all_rpv.to_csv(str(Path(base_dir) / "RPV_Results" / "V11"  / "rpv_results.csv"))
 
 
 if __name__ == "__main__":
