+++ src/Util/processing.py
@@ -2,18 +2,86 @@ import logging
 import time
 from datetime import datetime
 
+import numpy as np
 import polars as pl
 from tqdm import tqdm
 
-from src.Common.rpv import rpv_df_preprocess, rpv_fit
+from src.Common.rpv import rpv_fit
 
 
-def process_weekly_data(weeks_dics, band, debug=False, n_samples_bins=2000, sample_total_dataset=None):
+def rpv_df_preprocess(df, debug=False):
+    EPS = 1e-2  # Tolerance for floating point comparison
+
+    if "vx" in df.columns:
+        vx = df["vx"]
+        assert np.allclose(vx, df["xcam"] - df["Xw"], atol=EPS), "vx mismatch"
+    else:
+        vx = df["xcam"] - df["Xw"]
+        df = df.with_columns(pl.Series("vx", vx))
+
+    if "vy" in df.columns:
+        vy = df["vy"]
+        assert np.allclose(vy, df["ycam"] - df["Yw"], atol=EPS), "vy mismatch"
+    else:
+        vy = df["ycam"] - df["Yw"]
+        df = df.with_columns(pl.Series("vy", vy))
+
+    if "vz" in df.columns:
+        vz = df["vz"]
+        assert np.allclose(vz, df["delta_z"], atol=EPS), "vz mismatch"
+    else:
+        vz = df["delta_z"]
+        df = df.with_columns(pl.Series("vz", vz))
+
+    if "v_norm" in df.columns:
+        v_norm = df["v_norm"]
+        assert np.allclose(v_norm, np.sqrt(vx ** 2 + vy ** 2 + vz ** 2), atol=EPS), "v_norm mismatch"
+    else:
+        v_norm = np.sqrt(vx ** 2 + vy ** 2 + vz ** 2)
+        df = df.with_columns(pl.Series("v_norm", v_norm))
+
+    # Formulas as expressions
+    cos_vza = (df["vz"] / (df["v_norm"] + 1e-12)).clip(-1.0, 1.0)  # 1  guard /0 and range
+    vza_formula = np.degrees(np.arccos(cos_vza))
+    vaa_formula = np.degrees(np.arctan2(df["vx"], df["vy"])) % 360
+    sza_formula = 90 - df["sunelev"]
+
+    raa_formula = np.abs(df["saa"] - vaa_formula) % 360
+    raa_formula = np.where(raa_formula <= 180, raa_formula, 360 - raa_formula)
+
+    # indexes
+    ndvi_formula = (df["band5"] - df["band3"]) / (df["band5"] + df["band3"])
+    excess_green_formula = (2 * df["band2"] - df["band3"] - df["band1"])
+
+    Y = 0.16
+    osavi_formula = (1 + Y) * (df["band5"] - df["band3"]) / (df["band5"] + df["band3"] + Y)
+
+    # Check or create columns
+    for col, formula in [
+        ("vza", vza_formula),
+        ("vaa", vaa_formula),
+        ("sza", sza_formula),
+        ("NDVI", ndvi_formula),
+        ("raa", raa_formula),
+        ("excess_green", excess_green_formula),
+        ("osavi", osavi_formula),
+    ]:
+        if col in df.columns and debug == True:
+            if not np.allclose(df[col], formula, atol=EPS):
+                print(f"Column '{col}' values do not match formula!")
+                df = df.with_columns(pl.Series(col, formula))
+        else:
+            df = df.with_columns(pl.Series(col, formula))
+    return df
+
+
+def process_weekly_data(weeks_dics, band, debug=False, n_samples_bins=5000, sample_total_dataset=None , filter ={}):
     """
     Process RPV data for each week and return results as a Polars DataFrame
 
     Args:
         weeks_dics (dict): Dictionary with week IDs as keys and dataframes as values
+        filter (dict): Filter a column with a treshholdthe data before fitting the model.
 
     Returns:
         pl.DataFrame: A Polars DataFrame containing all RPV analysis results
@@ -53,6 +121,12 @@ def process_weekly_data(weeks_dics, band, debug=False, n_samples_bins=2000, samp
                         dg = dg.sample(sample_total_dataset)
                 dg = rpv_df_preprocess(dg, debug)
 
+
+                if filter:
+                    if filter["sign"] == ">":
+                        dg = dg.filter(pl.col(filter["column"]) > filter["threshold"] )
+                    if filter["sign"] == "<":
+                        dg = dg.filter(pl.col(filter["column"]) < filter["threshold"])
